Extract live map card rendering
